refactor: move active -> ../enabled
