+++ init.lua
@@ -2,4 +2,4 @@ require('nvim2k.options')
 require('nvim2k.keymaps')
 require('nvim2k.functions')
 require('nvim2k.autocmd')
-require('nvim2k.plugins.active')
+require('nvim2k.enabled')
